--- tunup/README.md
@@ -1,1 +1,0 @@
-keep tunnel from quasar to core-v4 up by periodically sending http request
